feat: new /api/v2/logs for local eth_getLog and eth_getFilterLogs
Old /api/log is implemented with serving the browser frontend in mind
when it was first impl in explorer server.
It has fairly sophisticated pagination system and does not support proper
blockHash field required by eth_getLog. I've avoided to modify the logic
code of /api/log to make this work purely because there might be some
legacy code that's expected as it is. The addition of new api is simple
because it's not trying to be abunch of different things.